+++ rl_zoo3/train.py
@@ -12,7 +12,7 @@ import torch as th
 from stable_baselines3.common.utils import set_random_seed
 
 # Register custom envs
-import rl_zoo3.import_envs  # noqa: F401 pytype: disable=import-error
+import rl_zoo3.import_envs  # noqa: F401
 from rl_zoo3.exp_manager import ExperimentManager
 from rl_zoo3.utils import ALGOS, StoreDict
 
@@ -164,7 +164,7 @@ def train() -> None:
         importlib.import_module(env_module)
 
     env_id = args.env
-    registered_envs = set(gym.envs.registry.keys())  # pytype: disable=module-attr
+    registered_envs = set(gym.envs.registry.keys())
 
     # If the environment is not found, suggest the closest match
     if env_id not in registered_envs: